+++ android/src/main/kotlin/com/snowplowanalytics/snowplow_tracker/readers/messages/EventMessageReader.kt
@@ -1,4 +1,4 @@
-// Copyright (c) 2022 Snowplow Analytics Ltd. All rights reserved.
+// Copyright (c) 2022-present Snowplow Analytics Ltd. All rights reserved.
 //
 // This program is licensed to you under the Apache License Version 2.0,
 // and you may not use this file except in compliance with the Apache License Version 2.0.
@@ -27,37 +27,49 @@ class EventMessageReader(val values: Map<String, Any>) {
 
     fun toStructuredWithContexts(): Structured {
         val structured = StructuredReader(eventData).toStructured()
-        contextsJsons?.let { structured.customContexts.addAll(it) }
+        contextsJsons?.let { structured.entities.addAll(it) }
         return structured
     }
 
     fun toSelfDescribingWithContexts(): SelfDescribing {
         val selfDescribing = SelfDescribingJsonReader(eventData).toSelfDescribing()
-        contextsJsons?.let { selfDescribing.customContexts.addAll(it) }
+        contextsJsons?.let { selfDescribing.entities.addAll(it) }
         return selfDescribing
     }
 
     fun toScreenViewWithContexts(): ScreenView {
         val screenView = ScreenViewReader(eventData).toScreenView()
-        contextsJsons?.let { screenView.customContexts.addAll(it) }
+        contextsJsons?.let { screenView.entities.addAll(it) }
         return screenView
+    }
+
+    fun toScrollChangedWithContexts(): ScrollChanged {
+        val scrollChanged = ScrollChangedReader(eventData).toScrollChanged()
+        contextsJsons?.let { scrollChanged.entities.addAll(it) }
+        return scrollChanged
+    }
+
+    fun toListItemViewWithContexts(): ListItemView {
+        val listItemView = ListItemViewReader(eventData).toListItemView()
+        contextsJsons?.let { listItemView.entities.addAll(it) }
+        return listItemView
     }
 
     fun toTimingWithContexts(): Timing {
         val timing = TimingReader(eventData).toTiming()
-        contextsJsons?.let { timing.customContexts.addAll(it) }
+        contextsJsons?.let { timing.entities.addAll(it) }
         return timing
     }
 
     fun toConsentGrantedWithContexts(): ConsentGranted {
         val consentGranted = ConsentGrantedReader(eventData).toConsentGranted()
-        contextsJsons?.let { consentGranted.customContexts.addAll(it) }
+        contextsJsons?.let { consentGranted.entities.addAll(it) }
         return consentGranted
     }
 
     fun toConsentWithdrawnWithContexts(): ConsentWithdrawn {
         val consentWithdrawn = ConsentWithdrawnReader(eventData).toConsentWithdrawn()
-        contextsJsons?.let { consentWithdrawn.customContexts.addAll(it) }
+        contextsJsons?.let { consentWithdrawn.entities.addAll(it) }
         return consentWithdrawn
     }
 }