+++ lib/src/web/readers/messages/event_message_reader.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2022 Snowplow Analytics Ltd. All rights reserved.
+// Copyright (c) 2022-present Snowplow Analytics Ltd. All rights reserved.
 //
 // This program is licensed to you under the Apache License Version 2.0,
 // and you may not use this file except in compliance with the Apache License Version 2.0.
@@ -19,6 +19,8 @@ import '../events/self_describing_reader.dart';
 import '../events/structured_reader.dart';
 import '../events/timing_reader.dart';
 import '../events/page_view_event_reader.dart';
+import '../events/scroll_changed_reader.dart';
+import '../events/list_item_view_reader.dart';
 
 @immutable
 class EventMessageReader {
@@ -26,6 +28,8 @@ class EventMessageReader {
   final StructuredReader? structured;
   final SelfDescribingReader? selfDescribing;
   final ScreenViewReader? screenView;
+  final ScrollChangedReader? scrollChanged;
+  final ListItemViewReader? listItemView;
   final TimingReader? timing;
   final PageViewEventReader? pageView;
   final ConsentGrantedReader? consentGranted;
@@ -37,6 +41,8 @@ class EventMessageReader {
       this.structured,
       this.selfDescribing,
       this.screenView,
+      this.scrollChanged,
+      this.listItemView,
       this.timing,
       this.consentGranted,
       this.consentWithdrawn,
@@ -48,6 +54,8 @@ class EventMessageReader {
         structured = StructuredReader(map['eventData']),
         selfDescribing = null,
         screenView = null,
+        scrollChanged = null,
+        listItemView = null,
         timing = null,
         consentGranted = null,
         consentWithdrawn = null,
@@ -60,6 +68,8 @@ class EventMessageReader {
         selfDescribing = SelfDescribingReader(map['eventData']),
         structured = null,
         screenView = null,
+        scrollChanged = null,
+        listItemView = null,
         timing = null,
         consentGranted = null,
         consentWithdrawn = null,
@@ -72,6 +82,36 @@ class EventMessageReader {
         selfDescribing = null,
         structured = null,
         screenView = ScreenViewReader(map['eventData']),
+        scrollChanged = null,
+        listItemView = null,
+        timing = null,
+        consentGranted = null,
+        consentWithdrawn = null,
+        pageView = null,
+        contexts =
+            map['contexts'] != null ? ContextsReader(map['contexts']) : null;
+
+  EventMessageReader.withScrollChanged(dynamic map)
+      : tracker = map['tracker'],
+        selfDescribing = null,
+        structured = null,
+        screenView = null,
+        scrollChanged = ScrollChangedReader(map['eventData']),
+        listItemView = null,
+        timing = null,
+        consentGranted = null,
+        consentWithdrawn = null,
+        pageView = null,
+        contexts =
+            map['contexts'] != null ? ContextsReader(map['contexts']) : null;
+
+  EventMessageReader.withListItemView(dynamic map)
+      : tracker = map['tracker'],
+        selfDescribing = null,
+        structured = null,
+        screenView = null,
+        scrollChanged = null,
+        listItemView = ListItemViewReader(map['eventData']),
         timing = null,
         consentGranted = null,
         consentWithdrawn = null,
@@ -84,6 +124,8 @@ class EventMessageReader {
         selfDescribing = null,
         structured = null,
         screenView = null,
+        scrollChanged = null,
+        listItemView = null,
         timing = TimingReader(map['eventData']),
         consentGranted = null,
         consentWithdrawn = null,
@@ -96,6 +138,8 @@ class EventMessageReader {
         selfDescribing = null,
         structured = null,
         screenView = null,
+        scrollChanged = null,
+        listItemView = null,
         timing = null,
         consentGranted = ConsentGrantedReader(map['eventData']),
         consentWithdrawn = null,
@@ -108,6 +152,8 @@ class EventMessageReader {
         selfDescribing = null,
         structured = null,
         screenView = null,
+        scrollChanged = null,
+        listItemView = null,
         timing = null,
         consentGranted = null,
         consentWithdrawn = ConsentWithdrawnReader(map['eventData']),
@@ -120,6 +166,8 @@ class EventMessageReader {
         selfDescribing = null,
         structured = null,
         screenView = null,
+        scrollChanged = null,
+        listItemView = null,
         timing = null,
         consentGranted = null,
         consentWithdrawn = null,
@@ -131,6 +179,8 @@ class EventMessageReader {
     if (structured != null) return structured!;
     if (selfDescribing != null) return selfDescribing!;
     if (screenView != null) return screenView!;
+    if (scrollChanged != null) return scrollChanged!;
+    if (listItemView != null) return listItemView!;
     if (timing != null) return timing!;
     if (consentGranted != null) return consentGranted!;
     if (pageView != null) return pageView!;