+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 0.6.0
+
+* Upgrade mobile trackers to version 6.0
+* Add support for mobile screen engagement tracking including list item view and scroll changed events (#43)
+* Add configuration to override platform context properties in the mobile context including the IDFA identifiers (#44)
+* Update SDK constraint to <4.0.0 (#42)
+* Fix publish action (#38)
+* Remove deprecated setMockMethodCallHandler from tests (#41)
+* Move the repository to the snowplow Github organization (#29)
+* Update copyright headers
+
 # 0.5.0
 
 *  Add lifecycle autotracking config option (#39)